edit: cookie 컴포넌트 삭제 및 통합
